benchmarks: add reproducible auto_clean memory benchmark
