+++ README.md
@@ -325,34 +325,22 @@ python benchmarks/benchmark_vs_pandas.py
 
 `benchmarks/generate_data.py` uses deterministic NumPy seeds, so every run creates the same `benchmarks/benchmark_1m.csv` tall input and `benchmarks/benchmark_wide.csv` wide input. The benchmark then executes three pandas runs and three arnio runs for each case, printing average wall-clock time from `time.perf_counter()` and peak Python allocation from `tracemalloc`. For cleaner comparisons, close other memory-heavy processes and run the script from the repository root after installing the same Python, pandas, NumPy, compiler, and arnio commit you want to compare.
 
-### 🧠 Auto Clean Memory Benchmark
-
-To measure the peak memory and execution time of the `auto_clean` pipeline specifically on realistic data sizes:
-
-```bash
-python benchmarks/benchmark.py --rows 100000
-```
-
-This script generates a reproducible synthetic dataset with mixed column types (strings, ints, floats, booleans, nulls, and duplicates) and measures:
-- `ar.read_csv` performance
-- `ar.auto_clean(mode="safe")` performance (low-risk cleanup like whitespace trimming)
-- `ar.auto_clean(mode="strict")` performance (includes type casting and deduplication)
-
-Options:
-- `--repeat N` runs each operation multiple times and reports average (and min/max range).
-- `--seed N` changes the deterministic dataset seed.
-- `--reuse-file` reuses an existing dataset file instead of regenerating it.
-
 Expected output format:
 
 ```text
-Operation                     Time(s)     Peak Py(MiB)
-------------------------------------------------------------
-ar.read_csv            0.042 (0.041-0.044)    4.52 (4.50-4.60)
-ar.auto_clean(safe)    0.012 (0.011-0.013)    0.15 (0.14-0.16)
-ar.auto_clean(strict)  0.035 (0.034-0.036)    1.20 (1.18-1.22)
-------------------------------------------------------------
-Total avg (Read+Strict)              0.077             4.52
+Tall CSV (1,000,000 rows x 12 columns)
+Metric                     pandas        arnio
+────────────────────────────────────────────
+Exec Time (avg)       4.73s         5.75s
+Peak RAM               211MB         212MB
+Speed: 0.8x | RAM: -1% reduction
+
+Wide CSV (5,000 rows x 256 columns)
+Metric                     pandas        arnio
+────────────────────────────────────────────
+Exec Time (avg)       ...s          ...s
+Peak RAM              ...MB         ...MB
+Speed: ...x | RAM: ...% reduction
 ```
 
 Small differences are expected across CPUs, operating systems, compilers, Python builds, and pandas/NumPy versions. If you share benchmark results in an issue or PR, include your OS, Python version, CPU model, pandas/NumPy versions, arnio commit, and the full command output so maintainers can compare like for like.
@@ -384,6 +372,41 @@ Small differences are expected across CPUs, operating systems, compilers, Python
 </tr>
 </table>
 
+<br>
+
+### 🧠 Auto Clean Memory Benchmark
+
+To measure the peak memory and execution time of the `auto_clean` pipeline using realistic dataset sizes:
+
+```bash
+python benchmarks/benchmark_auto_clean_memory.py --rows 100000
+```
+
+This script generates a reproducible synthetic dataset with mixed column types (strings, ints, floats, booleans, nulls, and duplicates) and measures:
+- `ar.read_csv` performance
+- `ar.auto_clean(mode="safe")` performance (low-risk cleanup like whitespace trimming)
+- `ar.auto_clean(mode="strict")` performance (includes type casting and deduplication)
+
+The dataset is regenerated deterministically unless `--reuse-file` is provided.
+Each `auto_clean` benchmark run reloads the dataset to avoid mutation or caching effects between runs.
+
+Options:
+- `--repeat N` runs each operation multiple times and reports average (and min/max range).
+- `--seed N` changes the deterministic dataset seed.
+- `--reuse-file` reuses an existing dataset file instead of regenerating it.
+- `--keep-file` keeps the generated CSV (otherwise it is removed at the end).
+
+Expected output format:
+
+```text
+Operation                    Time(s)     Peak Py(MiB)
+--------------------------------------------------------------------
+ar.read_csv           0.042 (0.041-0.044)    4.52 (4.50-4.60)
+ar.auto_clean(safe)   0.012 (0.011-0.013)    0.15 (0.14-0.16)
+ar.auto_clean(strict) 0.035 (0.034-0.036)    1.20 (1.18-1.22)
+--------------------------------------------------------------------
+Total avg (Read+Strict)       0.077             4.52
+```
 <br>
 
 ---
